+++ cardinal-components-block/src/main/java/dev/onyxstudios/cca/internal/block/StaticBlockComponentPlugin.java
@@ -24,11 +24,15 @@ package dev.onyxstudios.cca.internal.block;
 
 import dev.onyxstudios.cca.api.v3.block.BlockComponentFactoryRegistry;
 import dev.onyxstudios.cca.api.v3.block.BlockComponentInitializer;
-import dev.onyxstudios.cca.api.v3.component.*;
+import dev.onyxstudios.cca.api.v3.component.Component;
+import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
+import dev.onyxstudios.cca.api.v3.component.ComponentFactory;
+import dev.onyxstudios.cca.api.v3.component.ComponentKey;
+import dev.onyxstudios.cca.api.v3.component.ComponentProvider;
 import dev.onyxstudios.cca.api.v3.component.tick.ClientTickingComponent;
 import dev.onyxstudios.cca.api.v3.component.tick.ServerTickingComponent;
 import dev.onyxstudios.cca.internal.base.LazyDispatcher;
-import dev.onyxstudios.cca.internal.base.asm.CcaAsmHelper;
+import dev.onyxstudios.cca.internal.base.QualifiedComponentFactory;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentLoadingException;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentPluginBase;
 import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap;
@@ -38,16 +42,20 @@ import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.function.Predicate;
 
 public final class StaticBlockComponentPlugin extends LazyDispatcher implements BlockComponentFactoryRegistry {
     public static final StaticBlockComponentPlugin INSTANCE = new StaticBlockComponentPlugin();
 
     private static String getSuffix(Class<? extends BlockEntity> key) {
-        String simpleName = key.getSimpleName();
-        return String.format("BlockEntityImpl_%s_%s", simpleName, Integer.toHexString(key.getName().hashCode()));
+        return "BlockEntityImpl_%s_%s".formatted(key.getSimpleName(), Integer.toHexString(key.getName().hashCode()));
     }
 
     private StaticBlockComponentPlugin() {
@@ -55,9 +63,7 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
     }
 
     private final List<PredicatedComponentFactory<?>> dynamicFactories = new ArrayList<>();
-    private final Map<Class<? extends BlockEntity>, Map<ComponentKey<?>, Class<? extends Component>>> beComponentImpls = new HashMap<>();
-    private final Map<Class<? extends BlockEntity>, Map<ComponentKey<?>, ComponentFactory<?, ?>>> beComponentFactories = new Reference2ObjectOpenHashMap<>();
-    private final Map<Class<? extends BlockEntity>, Class<? extends ComponentContainer.Factory<BlockEntity>>> factoryClasses = new Reference2ObjectOpenHashMap<>();
+    private final Map<Class<? extends BlockEntity>, Map<ComponentKey<?>, QualifiedComponentFactory<ComponentFactory<? extends BlockEntity, ?>>>> beComponentFactories = new Reference2ObjectOpenHashMap<>();
     private final Set<Class<? extends BlockEntity>> clientTicking = new ReferenceOpenHashSet<>();
     private final Set<Class<? extends BlockEntity>> serverTicking = new ReferenceOpenHashSet<>();
 
@@ -81,69 +87,61 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
 
     public boolean requiresStaticFactory(Class<? extends BlockEntity> entityClass) {
         StaticBlockComponentPlugin.INSTANCE.ensureInitialized();
+
+        for (PredicatedComponentFactory<?> dynamicFactory : this.dynamicFactories) {
+            dynamicFactory.tryRegister(entityClass);
+        }
+
         return entityClass == BlockEntity.class || this.beComponentFactories.containsKey(entityClass);
     }
 
-    public Class<? extends ComponentContainer.Factory<BlockEntity>> spinDedicatedFactory(Class<? extends BlockEntity> key) {
+    public ComponentContainer.Factory<BlockEntity> buildDedicatedFactory(Class<? extends BlockEntity> entityClass) {
         StaticBlockComponentPlugin.INSTANCE.ensureInitialized();
 
-        // we need a cache as this method is called for a given class each time one of its subclasses is loaded.
-        return this.factoryClasses.computeIfAbsent(key, entityClass -> {
-            for (PredicatedComponentFactory<?> dynamicFactory : this.dynamicFactories) {
-                dynamicFactory.tryRegister(entityClass);
+        var compiled = new LinkedHashMap<>(this.beComponentFactories.getOrDefault(entityClass, Collections.emptyMap()));
+        Class<? extends BlockEntity> type = entityClass;
+
+        while (type != BlockEntity.class) {
+            type = type.getSuperclass().asSubclass(BlockEntity.class);
+            for (var e : this.beComponentFactories.getOrDefault(type, Collections.emptyMap()).entrySet()) {
+                compiled.putIfAbsent(e.getKey(), e.getValue());
+                if (ClientTickingComponent.class.isAssignableFrom(e.getValue().impl())) this.clientTicking.add(entityClass);
+                if (ServerTickingComponent.class.isAssignableFrom(e.getValue().impl())) this.serverTicking.add(entityClass);
             }
+        }
 
-            Map<ComponentKey<?>, ComponentFactory<?, ?>> compiled = new LinkedHashMap<>(this.beComponentFactories.getOrDefault(entityClass, Collections.emptyMap()));
-            Map<ComponentKey<?>, Class<? extends Component>> compiledImpls = new LinkedHashMap<>(this.beComponentImpls.getOrDefault(entityClass, Collections.emptyMap()));
-            Class<? extends BlockEntity> type = entityClass;
+        ComponentContainer.Factory.Builder<BlockEntity> builder = ComponentContainer.Factory.builder(BlockEntity.class)
+            .factoryNameSuffix(getSuffix(entityClass));
 
-            while (type != BlockEntity.class) {
-                type = type.getSuperclass().asSubclass(BlockEntity.class);
-                this.beComponentFactories.getOrDefault(type, Collections.emptyMap()).forEach(compiled::putIfAbsent);
-                for (Map.Entry<ComponentKey<?>, Class<? extends Component>> entry : this.beComponentImpls.getOrDefault(type, Collections.emptyMap()).entrySet()) {
-                    Class<? extends Component> impl = entry.getValue();
-                    if (ClientTickingComponent.class.isAssignableFrom(impl)) this.clientTicking.add(entityClass);
-                    if (ServerTickingComponent.class.isAssignableFrom(impl)) this.serverTicking.add(entityClass);
-                    compiledImpls.putIfAbsent(entry.getKey(), impl);
-                }
-            }
+        for (var entry : compiled.entrySet()) {
+            addToBuilder(builder, entry);
+        }
 
-            String implSuffix = getSuffix(entityClass);
+        return builder.build();
+    }
 
-            try {
-                Class<? extends ComponentContainer> containerCls = CcaAsmHelper.spinComponentContainer(
-                    ComponentFactory.class,
-                    compiled,
-                    compiledImpls,
-                    implSuffix
-                );
-
-                return StaticComponentPluginBase.spinContainerFactory(
-                    implSuffix,
-                    ComponentContainer.Factory.class,
-                    containerCls,
-                    entityClass
-                );
-            } catch (IOException e) {
-                throw new StaticComponentLoadingException("Failed to generate a dedicated component container for " + entityClass, e);
-            }
-        });
+    private <C extends Component> void addToBuilder(ComponentContainer.Factory.Builder<BlockEntity> builder, Map.Entry<ComponentKey<?>, QualifiedComponentFactory<ComponentFactory<? extends BlockEntity, ?>>> entry) {
+        @SuppressWarnings("unchecked") var key = (ComponentKey<C>) entry.getKey();
+        @SuppressWarnings("unchecked") var factory = (ComponentFactory<BlockEntity, C>) entry.getValue().factory();
+        @SuppressWarnings("unchecked") var impl = (Class<C>) entry.getValue().impl();
+        builder.component(key, impl, factory, entry.getValue().dependencies());
     }
 
     public <C extends Component, E extends BlockEntity> void registerFor(Class<E> target, ComponentKey<C> type, ComponentFactory<E, C> factory) {
         this.checkLoading(BlockComponentFactoryRegistry.class, "register");
-        this.register0(target, type, factory, type.getComponentClass());
+        this.register0(target, type, new QualifiedComponentFactory<>(factory, type.getComponentClass(), Set.of()));
     }
 
-    private <C extends Component, F extends C, E extends BlockEntity> void register0(Class<? extends E> target, ComponentKey<? super C> type, ComponentFactory<E, F> factory, Class<C> impl) {
-        Map<ComponentKey<?>, ComponentFactory<?, ?>> specializedMap = this.beComponentFactories.computeIfAbsent(target, t -> new LinkedHashMap<>());
-        ComponentFactory<?, ?> previousFactory = specializedMap.get(type);
+    private <C extends Component, F extends C, E extends BlockEntity> void register0(Class<? extends E> target, ComponentKey<? super C> type, QualifiedComponentFactory<ComponentFactory<E, F>> factory) {
+        var specializedMap = this.beComponentFactories.computeIfAbsent(target, t -> new LinkedHashMap<>());
+        var previousFactory = specializedMap.get(type);
+
         if (previousFactory != null) {
-            throw new StaticComponentLoadingException("Duplicate factory declarations for " + type.getId() + " on " + target + ": " + factory + " and " + previousFactory);
+            throw new StaticComponentLoadingException("Duplicate factory declarations for %s on %s: %s and %s".formatted(type.getId(), target, factory, previousFactory));
         }
-        ComponentFactory<E, Component> checked = entity -> Objects.requireNonNull(((ComponentFactory<E, ?>) factory).createComponent(entity), "Component factory " + factory + " for " + type.getId() + " returned null on " + entity.getClass().getSimpleName());
-        specializedMap.put(type, checked);
-        this.beComponentImpls.computeIfAbsent(target, t -> new LinkedHashMap<>()).put(type, impl);
+
+        @SuppressWarnings("unchecked") var factory1 = (QualifiedComponentFactory<ComponentFactory<? extends BlockEntity, ?>>) (QualifiedComponentFactory<?>) factory;
+        specializedMap.put(type, factory1);
     }
 
     @Override
@@ -162,19 +160,17 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
     private final class PredicatedComponentFactory<C extends Component> {
         private final Predicate<Class<? extends BlockEntity>> predicate;
         private final ComponentKey<? super C> type;
-        private final ComponentFactory<BlockEntity, C> factory;
-        private final Class<C> impl;
+        private final QualifiedComponentFactory<ComponentFactory<BlockEntity, C>> factory;
 
-        public PredicatedComponentFactory(Predicate<Class<? extends BlockEntity>> predicate, ComponentKey<? super C> type, ComponentFactory<BlockEntity, C> factory, Class<C> impl) {
+        public PredicatedComponentFactory(Predicate<Class<? extends BlockEntity>> predicate, ComponentKey<? super C> type, QualifiedComponentFactory<ComponentFactory<BlockEntity, C>> factory) {
             this.type = type;
             this.factory = factory;
             this.predicate = predicate;
-            this.impl = impl;
         }
 
         public void tryRegister(Class<? extends BlockEntity> clazz) {
             if (this.predicate.test(clazz)) {
-                StaticBlockComponentPlugin.this.register0(clazz, this.type, this.factory, this.impl);
+                StaticBlockComponentPlugin.this.register0(clazz, this.type, this.factory);
             }
         }
     }
@@ -182,12 +178,14 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
     private final class RegistrationImpl<C extends Component, E extends BlockEntity> implements Registration<C, E> {
         private final Class<E> target;
         private final ComponentKey<? super C> key;
+        private final Set<ComponentKey<?>> dependencies;
         private Class<C> componentClass;
         private Predicate<Class<? extends E>> test;
 
         RegistrationImpl(Class<E> target, ComponentKey<C> key) {
             this.target = target;
             this.componentClass = key.getComponentClass();
+            this.dependencies = new LinkedHashSet<>();
             this.test = null;
             this.key = key;
         }
@@ -195,6 +193,12 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
         @Override
         public Registration<C, E> filter(Predicate<Class<? extends E>> test) {
             this.test = this.test == null ? test : this.test.and(test);
+            return this;
+        }
+
+        @Override
+        public Registration<C, E> after(ComponentKey<?> dependency) {
+            this.dependencies.add(dependency);
             return this;
         }
 
@@ -212,15 +216,17 @@ public final class StaticBlockComponentPlugin extends LazyDispatcher implements 
                 StaticBlockComponentPlugin.this.register0(
                     this.target,
                     this.key,
-                    factory,
-                    this.componentClass
+                    new QualifiedComponentFactory<>(factory, this.componentClass, this.dependencies)
                 );
             } else {
                 StaticBlockComponentPlugin.this.dynamicFactories.add(new PredicatedComponentFactory<>(
                     c -> this.target.isAssignableFrom(c) && this.test.test(c.asSubclass(this.target)),
                     this.key,
-                    entity -> factory.createComponent(this.target.cast(entity)),
-                    this.componentClass
+                    new QualifiedComponentFactory<>(
+                        entity -> factory.createComponent(this.target.cast(entity)),
+                        this.componentClass,
+                        this.dependencies
+                    )
                 ));
             }
         }