+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/api/v3/util/provider/package-info.java
@@ -25,6 +25,7 @@
  */
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
+@Deprecated(forRemoval = true)
 package dev.onyxstudios.cca.api.v3.util.provider;
 
 import dev.onyxstudios.cca.api.v3.util.MethodsReturnNonnullByDefault;